Update Rust bindings to make the package publishable (#19)
* Move cardano code to its own folder

* Regenerate with new cddl-codegen

* work on json codegen

* Undo changes to Plutus.json

* cargo fmt

* Update github action cwd

* Fix git workflow configs

* Fix cargo fmt in github actions

* Fix clippy error

* Add stronger git hook

* cargo fmt

* Fix publish-helper script

* Add NPM packages to the README

* cargo fmt

+++ cardano/sdk/rust/examples/contract_deploy.rs
@@ -13,6 +13,7 @@ use cml_chain::transaction::{
     DatumOption, RequiredSigners, Transaction, TransactionInput, TransactionOutput,
     TransactionWitnessSet,
 };
+use cml_chain::utils::BigInt;
 use reqwest::header::{HeaderMap, HeaderValue};
 use reqwest::{header, Client, StatusCode};
 use serde::de::Error;
@@ -27,7 +28,7 @@ use std::path::PathBuf;
 use std::str::FromStr;
 use std::time::UNIX_EPOCH;
 
-use cardano_projected_nft_sdk::{OutRef, Owner, Redeem, State, Status};
+use cardano_projected_nft::{OutRef, Owner, Redeem, State, Status};
 use cml_chain::builders::redeemer_builder::RedeemerWitnessKey;
 use cml_chain::builders::tx_builder::{
     ChangeSelectionAlgo, TransactionBuilder, TransactionBuilderConfigBuilder,
@@ -683,7 +684,7 @@ async fn handle_unlock(
 
     let validity = ttl_by_posix(now - 100);
     let ttl = ttl_by_posix(now + 100);
-    let for_how_long = (now + 400) * 1000;
+    let for_how_long = BigInt::from((now + 400) * 1000);
 
     let new_datum = match config.control_nft.clone() {
         None => State {
